ai-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy==1.25.2
pandas==2.1.0
spacy==3.6.1
PyPDF2==3.0.1
docx2txt==0.8
pytest==7.4.2
black==23.7.0
flake8==6.1.0
mypy==1.5.1
httpx==0.24.1
types-requests==2.31.0.2

# Download spaCy models with:
# python -m spacy download en_core_web_md
# python -m spacy download en_core_web_sm

gunicorn==21.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
celery==5.3.4
redis==5.0.0
aiohttp==3.8.5 